add : MockList.swift 파일 생성 및 코드 이동
MockList 타입 CalculatorItemQueueTests.swift ->  MockList.swift 이동
