+++ Calculator/CalculatorTests/CalculatorItemQueueTests.swift
@@ -76,27 +76,4 @@ class CalculatorItemQueueTests: XCTestCase {
         XCTAssertEqual(testData.count, sut.count)
     }
 
-
-}
-
-class MockList<T>: List {
-    
-    var mockData: [T] = []
-    
-    var isEmpty: Bool { mockData.isEmpty }
-    
-    var count: Int { mockData.count }
-    
-    func append(_ data: T) {
-        mockData.append(data)
-    }
-    
-    func removeFirst() -> T? {
-        return mockData.remove(at: 0)
-    }
-    
-    func removeAll() {
-        mockData.removeAll()
-    }
-
 }
